Feature: Add build-date query
